requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
PyPDF2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai